Separate out Registry and clean up the api
