+++ lib/index.js
@@ -1,3 +1,3 @@
-import Persona from './persona'
-
-export { Persona }
+import { PublicPersona, Persona } from './persona'
+import Registry from './registry'
+export { Registry, Persona, PublicPersona }
