+++ lib/persona.js
@@ -1,86 +1,4 @@
-import uportRegistry from 'uport-registry'
 import { decodeToken } from 'jsontokens'
-
-// consensysnet registry address, default for now
-const DEFAULT_REGISTRY_ADDRESS = '0xb9C1598e24650437a3055F7f66AC1820c419a679'
-
-
-
-// TODO return schema.org specific key values with prefix @ ? or filter them?
-
-// TODO wording: uportID or address
-// TODO profile or persona? wording for the represenation of identity/profile in here
-// TODO not sure if or how address (for a profile) should be passed around or if it should even, maybe optional
-// TODO decide what functionality to expose, and reflect that in the documents
-// TODO don't add type value for attestations
-// TODO when creating two indexed profiles don't do twice the work, for example don't parse tokens twice.
-
-// const personaNetwork = new Persona.Network(ipfs, web3Provider, registryAddress)
-// const personaProfile =  personaNetwork.getProfile(address)  //promise, can return raw data or personaProfile object
-
-
-/** Class representing a uPort Registry. */
- class Registry {
-
-     /**
-      *  Class constructor.
-      *  Creates a new Registry object. The registryAddress is an optional argument and if not specified will be at the moment set to the default ropsten network uport-registry.
-      *
-      *  @memberof Registry#
-      *  @method          constructor
-      *  @param           {String}         ipfsProvider                                                        an ipfs provider
-      *  @param           {String}         web3Provider                                                        web3 provider
-      *  @param           {String}         [registryAddress='0xb9C1598e24650437a3055F7f66AC1820c419a679']      the uport-registry address to use.
-      *  @return          {Object}         self
-      */
-      constructor (ipfs, web3Provider, registryAddress) {
-        this.uportRegistry = uportRegistry
-
-        if (ipfs && web3Provider) {
-          this.uportRegistry.setIpfsProvider(ipfs)
-          this.uportRegistry.setWeb3Provider(web3Provider)
-        }
-
-        this.registryAddress = registryAddress || DEFAULT_REGISTRY_ADDRESS
-      }
-
-
-      /**
-       *  Gets the public profile JSON object stored in IPFS for the given address.
-       *
-       *  @memberof Registry#
-       *  @method           getData
-       *  @return           {Promise<JSON, Error>}            A promise that returns the JSON object stored in IPFS for the given address
-       */
-      getData(address) {
-          return this.uportRegistry.getAttributes(this.registryAddress, address)
-      }
-
-      /**
-       *  Gets the the data stored in IPFS for the given object and creates a PublicPersona object.
-       *
-       *  @memberof Registry#
-       *  @method           getProfile
-       *  @return           {Promise<PublicPersona, Error>}            A promise that returns a new PublicPersona object.
-       */
-      getProfile(address) {
-        return new Promise((resolve, reject) => {
-          this.getData(address)
-            .then(profile => resolve(new PublicPersona(profile, address)))
-        })
-      }
-
-      /**
-       *  Gets the data stored in IPFS for an array of given addresses and creates an array of PublicPersona objects.
-       *
-       *  @memberof Registry#
-       *  @method           getProfile
-       *  @return           {Promise<PublicPersona, Error>}            A promise that returns an array of new PublicPersona objects.
-       */
-      getProfiles(addresses) {
-        return Promise.all(addresses.map(address => getProfile(address)))
-      }
-}
 
 // May reduce this code with an abstract class
 
@@ -587,4 +505,4 @@ const decodeClaim = (token) => decodeToken(token)
 
 // TODO Claim Utilities
 
-module.exports = { Registry, PublicPersona, Persona }
+module.exports = { PublicPersona, Persona }
